+++ test/stories/index.js
@@ -4,6 +4,7 @@
 import button from "./button.stories.js";
 import segmented from "./segmented.stories.js";
 import toggle from "./toggle.stories.js";
+import toggleGroup from "./toggle-group.stories.js";
 // Layout
 import divider from "./divider.stories.js";
 import space from "./space.stories.js";
@@ -65,7 +66,7 @@ import skeleton from "./skeleton.stories.js";
 import result from "./result.stories.js";
 
 export const STORIES = [
-  button, segmented, toggle,
+  button, segmented, toggle, toggleGroup,
   divider, space, flex, grid, layout, splitter, titlebar,
   input, textarea, number, select, combobox, checkbox, switchStory, radioGroup, slider, date, time, color, rate, form,
   tabs, breadcrumb, pagination, steps, nav, menubar,